requirements.txt
tabulate==0.7.5
futures==3.0.5
pyyaml==3.12
requests==2.11.1
python-keystoneclient==3.6.0
python-novaclient==6.0.0
python-neutronclient==6.0.0
python-cinderclient==1.9.0
datadog==0.14.0
